Make verify_prehashed() an associated method of the generic VerifyingKey
